requirements.txt
apache-airflow-providers-datadog==1!2.0.1